Crate oak_functions_wasm

Wasm business logic provider based on Wasmi.

Structs

UserState

UserState holds the user request bytes and response bytes for a particular execution of an Oak Wasm module. The UserState also holds a reference to the logger and the enabled extensions.

WasmHandler

Constants

ALLOC_FUNCTION_NAME

Fixed name of the function to allocate memory. Every Oak Wasm module must provide this function.

MAIN_FUNCTION_NAME

Fixed name of the function to start a Wasm. Every Oak Wasm module must provide this function.

MEMORY_NAME

The name of the memory every Oak Wasm module has.

Type Definitions

AbiExtensionHandle

Type alias for the ExtensionHandle type, which must be cast into a ExtensionHandle.

AbiPointer

Type aliases for positions in Wasm linear memory. Any future 64-bit version of Wasm would use different types.

AbiPointerOffset

Type aliases for offsets in Wasm linear memory.