Crate oak_functions_wasm

Wasm business logic provider based on Wasmi.

Structs

WasmHandler

WasmState

WasmState holds runtime values for a particular execution instance of Wasm, handling a single user request. The methods here correspond to the ABI host functions that allow the Wasm module to exchange the request and the response with the Oak functions server. These functions translate values between Wasm linear memory and Rust types.

Constants

ABI_USIZE

Wasm type identifier for position/offset values in linear memory. Any future 64-bit version of Wasm would use a different value.

Type Definitions

AbiExtensionHandle

AbiPointer

AbiPointerOffset